Image linking disable doesn't work
It doesnt work, HTML also doesnt work, github would just link them back to the damn thing. people will yoink the sprites whenever we want them to or not,

Sorry flin

well the wiki is open source anyway, so we're still screwed even if I managed to do it anyway.